Rename ApacheUniqueIdProvider to ServerRequestIdProvider
