+++ spec/Provider/RequestId/ServerRequestIdProviderSpec.php
@@ -1,16 +1,21 @@
 <?php
 
-namespace spec\TheRedDot\MonologExtraBundle\Provider\Uid;
+namespace spec\TheRedDot\MonologExtraBundle\Provider\RequestId;
 
-use TheRedDot\MonologExtraBundle\Provider\RequestId\ApacheUniqueIdProvider;
+use TheRedDot\MonologExtraBundle\Provider\RequestId\ServerRequestIdProvider;
 use TheRedDot\MonologExtraBundle\Provider\RequestId\RequestIdProviderInterface;
 use PhpSpec\ObjectBehavior;
 
-class ApacheUniqueIdProviderSpec extends ObjectBehavior
+class ServerRequestIdProviderSpec extends ObjectBehavior
 {
+    public function let()
+    {
+        $this->beConstructedWith('UNIQUE_ID');
+    }
+
     public function it_is_initializable()
     {
-        $this->shouldHaveType(ApacheUniqueIdProvider::class);
+        $this->shouldHaveType(ServerRequestIdProvider::class);
     }
 
     public function it_implements_request_id_provider_interface()
@@ -31,7 +36,7 @@ class ApacheUniqueIdProviderSpec extends ObjectBehavior
 
     public function it_returns_request_id_from_server_data()
     {
-        $this->beConstructedWith(['UNIQUE_ID' => 'sqdfjhqsodukfhqsdui']);
+        $this->beConstructedWith('UNIQUE_ID', ['UNIQUE_ID' => 'sqdfjhqsodukfhqsdui']);
         $this->getRequestId()->shouldReturn('sqdfjhqsodukfhqsdui');
     }
 }
