+++ src/Provider/RequestId/ServerRequestIdProvider.php
@@ -2,14 +2,14 @@
 
 namespace TheRedDot\MonologExtraBundle\Provider\RequestId;
 
-final class ApacheUniqueIdProvider implements RequestIdProviderInterface
+final class ServerRequestIdProvider implements RequestIdProviderInterface
 {
     /**
      * @var string
      */
     private $requestId;
 
-    public function __construct(array $serverData = null)
+    public function __construct(string $serverField, array $serverData = null)
     {
         $requestId = uniqid();
 
@@ -17,8 +17,8 @@ final class ApacheUniqueIdProvider implements RequestIdProviderInterface
             $serverData = &$_SERVER;
         }
 
-        if (isset($serverData['UNIQUE_ID'])) {
-            $requestId = $serverData['UNIQUE_ID'];
+        if (isset($serverData[$serverField])) {
+            $requestId = $serverData[$serverField];
         }
 
         $this->requestId = $requestId;
